fix: remove supervisor and edit restart directly

+++ scripts/restart.sh
@@ -1,4 +1,15 @@
+#!/bin/bash
 source /home/ubuntu/.bashrc
-killall -9 java
 export JAVA_HOME=/home/ubuntu/.sdkman/candidates/java/current
-export $(grep -v '^#' .env | xargs) && ./app &
+killall -9 java
+while true
+do
+	if [[ $(curl localhost:8080/health -m 5 -s) == "OK" ]]; then
+		echo "API is up."
+	else
+		echo "API is down. Restarting..."
+		killall -9 java
+		export $(grep -v '^#' .env | xargs) && ./app &
+	fi
+	sleep 5
+done
